README.md
Automated-Attendance-face-recognition-using-mask
To implement student attendance system based on face recognition with mask and without mask of webcam’s images. Face recognition is a biometric system used for human identification and is widely used . Now face mask is an inevitable thing in our lifestyle to ensure our safety . Hence we came up with an idea of face recognition with mask and not compromising safety to slightest.